move comments logic to separate file
